+++ engines/redis/index.ts
@@ -276,7 +276,7 @@ function patchRedisConfig(
     `pidfile ${normalizePathForRedis(options.pidFile)}`,
   )
   if (options.bindAddress !== undefined) {
-    config = config.replace(/^bind .+/m, `bind ${options.bindAddress}`)
+    config = config.replace(/^bind\s+.*$/m, `bind ${options.bindAddress}`)
   }
   if (options.daemonize !== undefined) {
     config = config.replace(